[ros2][camera control] Change threading model for cameras
